audio_hal: a2dp memcpy_to_i32_from_i16 frames error [1/1]
PD#SWPL-58353

Problem:
[Boreal][b/193858377] Bad audio quality on Sony WH-1000XM3 bluetooth headset

Solution:
1.Reason:
1). Upmix to 32bit form 16bit input frames param error.
2.Solution:
1). Correct the parameter

Verify:
Verrify by Android P + T982

+++ audio_hal/a2dp_hal.h
@@ -46,7 +46,7 @@ typedef struct {
     void * vir_buf_handle;
     uint64_t last_write_time;
     char * buff_conv_format;
-    int buff_size_conv_format;
+    size_t buff_size_conv_format;
     uint64_t mute_time;
 } aml_a2dp_hal;
